move mapping output specifications to a mixin
