+++ src/cell_type_mapper/schemas/mixins.py
@@ -2,6 +2,9 @@ import argschema
 from marshmallow import post_load
 import pathlib
 
+from cell_type_mapper.utils.anndata_utils import (
+    does_obsm_have_key)
+
 
 class PrecomputedStatsPathMixin(object):
 
@@ -97,4 +100,76 @@ class OutFileWithClobberMixin(object):
                 "already exists. To overwrite, run with clobber=True"
             )
             raise RuntimeError(msg)
+        return data
+
+
+class OutputDstForSearchMixin(object):
+
+    extended_result_path = argschema.fields.OutputFile(
+        required=False,
+        default=None,
+        allow_none=True,
+        description="Path to JSON file where extended results "
+        "will be saved.")
+
+    csv_result_path = argschema.fields.OutputFile(
+        required=False,
+        default=None,
+        allow_none=True,
+        description="Path to CSV file where output file will be "
+        "written (if None, no CSV will be produced).")
+
+    summary_metadata_path = argschema.fields.OutputFile(
+        required=False,
+        default=None,
+        allow_none=True,
+        description=(
+            "If not None, the path to a JSON file where summary "
+            "metadata (e.g. number of mapped genes and number "
+            "of mapped cells) will be stored")
+        )
+
+    obsm_key = argschema.fields.String(
+        required=False,
+        default=None,
+        allow_none=True,
+        description="If not None, save the results of the "
+        "mapping in query_path.obsm under this key")
+
+    obsm_clobber = argschema.fields.Boolean(
+        required=False,
+        default=False,
+        allow_none=False,
+        description="If True, allow the code to overwrite an "
+        "existing element in query_path.obsm")
+
+    @post_load
+    def check_obsm_key(self, data, **kwargs):
+        """
+        If obsm_key is not None, make sure that key has not already
+        been assigned in query_path.obsm
+        """
+        if data['obsm_key'] is None:
+            return data
+
+        if does_obsm_have_key(data['query_path'], data['obsm_key']):
+            if not data['obsm_clobber']:
+                msg = (f"obsm in {data['query_path']} already has key "
+                       f"{data['obsm_key']}; to overwrite, set obsm_clobber "
+                       "to True.")
+                raise RuntimeError(msg)
+
+        return data
+
+    @post_load
+    def check_result_dst(self, data, **kwargs):
+        """
+        Make sure that there is somewhere, either extended_result_path
+        or obsm_key, where we can store the extended results.
+        """
+        if data['extended_result_path'] is None:
+            if data['obsm_key'] is None:
+                msg = ("You must specify at least one of extended_result_path "
+                       "and/or obsm_key")
+                raise RuntimeError(msg)
         return data
